a
asd

+++ Document-Approval-Tracking-and-Mngt-System/DATMS/admin/changePassword/Pitik.php
@@ -1,5 +1,5 @@
 <?php session_start();
-  if ($_GET["Pitik"] == $_SESSION["Confirmed"]){ $_SESSION["conpass"] = true;
+  if (isset($_SESSION["Confirmed"])  && $_GET["Pitik"] == $_SESSION["Confirmed"]){ $_SESSION["conpass"] = true;
     echo '<small  class="form-text text-muted" > <span style="color:green"> <i class="bi bi-check-circle"></i> New passsword matched</span></small>';
   }else{
     echo '<small  class="form-text text-muted" > <span style="color:red"> <i class="bi bi-exclamation-circle"></i> New password and confirm password "doesnt match" Or "dont match".</span></small>';
